alexnet_batchnorm.py
The script is the realization of command oriented style of AlexNet. It is similar to the original style 
of AlexNet script written by Michael Guerzhoy on 2017 and keeps the computing parameters(60M+)
no changed.  In consideration of the actual inputs with 1000 classes, we set the same class numbers. 
The model is distinguished with Batch Normalization.