Update client_one.py
Set up the GPU in the condition of allocation exceeds system memory. For example, you will have the problem
if you you have 16GB memory. It avoids the the error and sudden stop of the runtime with the reminding message: Could not create cuDNN handle. Please see the added lines of code as follows. 

gpus = tf.config.experimental.list_physical_devices('GPU')
for gpu in gpus:
    tf.config.experimental.set_memory_growth(gpu, True)