Update client_one.py
Set up the GPU in the condition of allocation exceeds system memory.  For example, if you have only 16 GB memory,  the computer will have  a sudden stop of the runtime with the reminding message: Could not create cuDNN handle. The added code avoids the issue.  Please see the added lines of code as follows.

gpus = tf.config.experimental.list_physical_devices('GPU')
for gpu in gpus:
    tf.config.experimental.set_memory_growth(gpu, True)