Build the client for different scenarios
Split the client application from the original single script, build the AlexNet model for TensorFlow 2.0 based on the script written by the AlexNet team and add necessary lines of code for better readability, accessibility and usability. In addition, Mike corrects both the logical and code of line errors of the original script. Developers need to install the related libraries addressed in the readme file