Comply with TensorFlow 2.2 and Keras 2.4.3
The simplified client only includes one scenario of run_experiment(n, large_data_set=False, generator=False) as the last line of code and delete 100 lines of code in order to streamline the client. 

Set up the GPU in the condition of allocation exceeds system memory.  For example, if you have only 16 GB memory,  the computer will have  a sudden stop of the runtime with the reminding message: Could not create cuDNN handle. The added code avoids the runtime problem.

Add the one-hot lines of code into the preprocessing od train images  in order to comply with TensotFlow 2.2 and Keras 2.4.3

    train_images = np.array(train_images)
    train_labels = np.array(train_labels)
    train_labels = utils.to_categorical(train_labels)